Switch to regular JS action
