--- .dockerignore
@@ -1,7 +1,0 @@
-.idea
-tests
-.editorconfig
-.gitattributes
-.gitignore
-action.yml
-Dockerfile
