+++ .editorconfig
@@ -5,7 +5,5 @@ charset = utf-8
 end_of_line = lf
 insert_final_newline = true
 max_line_length = 120
-
-[{*.json,.babelrc,.eslintrc,.prettierrc,jest.config,*.yaml,*.yml}]
 indent_size = 2
 indent_style = space
