--- .github/workflows/build-and-push.yaml
@@ -1,64 +1,0 @@
-name: Build and Push Docker Image
-on:
-  push:
-    branches:
-      - master
-
-permissions:
-  contents: read
-  packages: write
-  attestations: write
-  id-token: write
-
-env:
-  REGISTRY: ghcr.io
-  IMAGE_NAME: taservers/vcpkg-cache
-
-jobs:
-  build-and-push:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v4
-
-      - name: Parse CHANGELOG.md
-        uses: mindsers/changelog-reader-action@v2
-        id: changelog
-        with:
-          validation_level: warn
-
-      - name: Login to Docker Hub
-        uses: docker/login-action@v3
-        with:
-          registry: ${{ env.REGISTRY }}
-          username: ${{ github.actor }}
-          password: ${{ secrets.GITHUB_TOKEN }}
-
-      - name: Set up Docker Buildx
-        uses: docker/setup-buildx-action@v3
-
-      - name: Extract metadata (tags, labels) for Docker
-        id: meta
-        uses: docker/metadata-action@v5
-        with:
-          images: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
-          tags: |
-            type=semver,pattern={{version}},value=v${{ steps.changelog.outputs.version }}
-            type=semver,pattern={{major}}.{{minor}},value=v${{ steps.changelog.outputs.version }}
-            type=semver,pattern={{major}},value=v${{ steps.changelog.outputs.version }}
-
-      - name: Build and push
-        id: push
-        uses: docker/build-push-action@v6
-        with:
-          context: .
-          file: Dockerfile
-          push: true
-          tags: ${{ steps.meta.outputs.tags }}
-          labels: ${{ steps.meta.outputs.labels }}
-
-      - name: Generate artifact attestation
-        uses: actions/attest-build-provenance@v2
-        with:
-          subject-name: ${{ env.REGISTRY }}/${{ env.IMAGE_NAME }}
-          subject-digest: ${{ steps.push.outputs.digest }}
-          push-to-registry: true
