+++ CHANGELOG.md
@@ -14,6 +14,12 @@ The valid change types are:
 - `Fixed` for any bug fixes.
 - `Security` in case of vulnerabilities.
 
+## [2.0.0] - [todo](todo)
+
+### Changed
+
+- **Breaking:** Converted to a regular Node.js action for compatibility with non-Linux runners
+
 ## [1.0.3] - [No PR](#)
 
 ### Fixed
