--- restore.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-deno run --allow-all /app/src/main.ts restore
